Fix the parameter erroor

+++ src/com/DocSystem/controller/ReposController.java
@@ -445,7 +445,8 @@ public class ReposController extends BaseController{
 	
 	/****************   get Repository Menu so that we can touch the docId******************/
 	@RequestMapping("/getReposInitMenu.do")
-	public void getReposInitMenu(Integer reposId,Integer docId, String parentPath, String docName, HttpSession session,HttpServletRequest request,HttpServletResponse response){
+	public void getReposInitMenu(Integer reposId,Long docId, String parentPath, String docName, HttpSession session,HttpServletRequest request,HttpServletResponse response)
+	{
 		System.out.println("getReposInitMenu reposId: " + reposId + " docId: " + docId + " parentPath:" + parentPath + " docName:" + docName);
 		
 		ReturnAjax rt = new ReturnAjax();
@@ -482,7 +483,7 @@ public class ReposController extends BaseController{
 		List <Doc> docList = null;
 		if(docId == null || docId == 0)
 		{
-			docId = 0;
+			docId = 0L;
 			docList = getAccessableSubDocList(repos, (long) 0, "", "", rootDocAuth, docAuthHashMap, rt);
 		}
 		else
